fix builds for android and non-x86_64 architectures
Android doesn't have a threads library. All pthread_* functions are
implemented in Bionic libc. This is one of the fundamental difference
between Android and Linux distributions and hence CMakeLists.txt was
patched not to look for a threads library on Android.

-march=native flag has been removed in CMakeLists.txt, this flag is
unsupported during cross compilations and also I feel that this flag
should always be passed using CXXFLAGS environment variables to CMake.

Includes for immintrin.h has been disabled for non-x86_64 and non-i686
platforms as this header is only supposed to be used on these platforms.

C language has been passed to project in CMakeLists.txt as it is now
required either for newer CMake versions (2.23+)